requirements.txt
absl-py==1.4.0
astunparse==1.6.3
attrs @ file:///croot/attrs_1668696182826/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.0.1
cycler==0.11.0
debugpy==1.6.6
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908585034/work
fastjsonschema @ file:///opt/conda/conda-bld/python-fastjsonschema_1661371079312/work
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
gast==0.3.3
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==2.10.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
ipykernel==6.12.1
ipython==7.31.1
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
jedi==0.18.2
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.2.0
jsonschema @ file:///croot/jsonschema_1676558650973/work
jupyter==1.0.0
jupyter-console==5.2.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-tuner==1.3.0
kiwisolver==1.4.4
kt-legacy==1.0.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.2
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
nb-conda-kernels @ file:///tmp/build/80754af9/nb_conda_kernels_1606775986548/work
nbclient==0.5.13
nbconvert==6.4.3
nbformat @ file:///croot/nbformat_1670352325207/work
nest-asyncio==1.5.6
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1628216317882/work
numpy==1.18.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.2.4
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
pkgutil_resolve_name @ file:///opt/conda/conda-bld/pkgutil-resolve-name_1661463321198/work
ply==3.11
prometheus-client @ file:///tmp/abs_d3zeliano1/croots/recipe/prometheus_client_1659455100375/work
prompt-toolkit==1.0.15
protobuf==3.20.3
psutil==5.9.4
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636098896055/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole @ file:///croot/qtconsole_1674008428467/work
QtPy @ file:///opt/conda/conda-bld/qtpy_1662014892439/work
rdkit-pypi==2021.3.1.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.4.1
seaborn==0.11.2
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
simplegeneric==0.8.1
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
terminado @ file:///croot/terminado_1671751832461/work
testpath @ file:///opt/conda/conda-bld/testpath_1655908557405/work
threadpoolctl==3.1.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
traitlets==5.9.0
typeguard==3.0.1
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension @ file:///tmp/build/80754af9/widgetsnbextension_1645009354203/work
wrapt==1.12.1
xlrd==1.2.0
zipp==3.15.0
